Move messaging-contacts-integration to features/messaging module
Relocate cross-domain bridge from orphaned src/app to features/messaging,
where the owning domain lives. Renamed to messaging-contacts-bridge.js
to signal its integration purpose.

Also removes obsolete src/app directory and 'app' type from eslint
boundaries config.
